Rename Lwjgl backend files to Gdx
